databasewebsite/requirements.txt
google-genai
supabase
python-dotenv
requests
